Refactor reusable components into separate files
